+++ src/core/server/sub/mod.rs
@@ -8,6 +8,7 @@ mod playlist;
 mod playlist_jump;
 mod pop_playlist;
 mod pop_set_playlist;
+mod remove_from_playlist;
 mod reorder_playlist_stack;
 mod set_all;
 mod set_playlist;
@@ -15,6 +16,6 @@ mod set_playlist;
 pub use self::{
     config::*, insert_into_playlist::*, library::*, new_server::*,
     play_tmp::*, player::*, playlist::*, playlist_jump::*, pop_playlist::*,
-    pop_set_playlist::*, reorder_playlist_stack::*, set_all::*,
-    set_playlist::*,
+    pop_set_playlist::*, remove_from_playlist::*, reorder_playlist_stack::*,
+    set_all::*, set_playlist::*,
 };